Refactor token code
- Rename file and structs to make their purpose more obvious
- Eliminate global tokenServer variable
